requirements.txt
tkinter
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2
python-dateutil==2.8.2
cryptography==41.0.8
bcrypt==4.1.2
tkcalendar==1.6.1
babel==2.13.1
arabic-reshaper==3.0.0
python-bidi==0.4.2